+++ Nutrons2014/src/com/nutrons/aerialassist/commands/ExampleCommand.java
@@ -1,16 +1,15 @@
 
-package edu.wpi.first.wpilibj.templates.commands.drivetrain;
-
-import edu.wpi.first.wpilibj.templates.commands.CommandBase;
+package com.nutrons.aerialassist.commands;
 
 /**
  *
- * @author camilo
+ * @author bradmiller
  */
-public class DTManualTankCmd extends CommandBase {
+public class ExampleCommand extends CommandBase {
 
-    public DTManualTankCmd() {
-        
+    public ExampleCommand() {
+        // Use requires() here to declare subsystem dependencies
+        // eg. requires(chassis);
     }
 
     // Called just before this Command runs the first time
@@ -19,7 +18,6 @@ public class DTManualTankCmd extends CommandBase {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-        dt.driveLR(oi.getDriveLeft(), oi.getDriveRight());
     }
 
     // Make this return true when this Command no longer needs to run execute()
