+++ Nutrons2014/src/com/nutrons/aerialassist/commands/drivetrain/DTManualTankCmd.java
@@ -1,15 +1,16 @@
 
-package edu.wpi.first.wpilibj.templates.commands;
+package com.nutrons.aerialassist.commands.drivetrain;
+
+import com.nutrons.aerialassist.commands.CommandBase;
 
 /**
  *
- * @author bradmiller
+ * @author camilo
  */
-public class ExampleCommand extends CommandBase {
+public class DTManualTankCmd extends CommandBase {
 
-    public ExampleCommand() {
-        // Use requires() here to declare subsystem dependencies
-        // eg. requires(chassis);
+    public DTManualTankCmd() {
+        
     }
 
     // Called just before this Command runs the first time
@@ -18,6 +19,7 @@ public class ExampleCommand extends CommandBase {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
+        dt.driveLR(oi.getDriveLeft(), oi.getDriveRight());
     }
 
     // Make this return true when this Command no longer needs to run execute()
