--- Nutrons2014/src/edu/wpi/first/wpilibj/templates/RobotMap.java
@@ -1,30 +1,0 @@
-package edu.wpi.first.wpilibj.templates;
-
-/**
- * The RobotMap is a mapping from the ports sensors and actuators are wired into
- * to a variable name. This provides flexibility changing wiring, makes checking
- * the wiring easier and significantly reduces the number of magic numbers
- * floating around.
- */
-public class RobotMap {
-    // For example to map the left and right motors, you could define the
-    // following variables to use with your drivetrain subsystem.
-    // public static final int leftMotor = 1;
-    // public static final int rightMotor = 2;
-    
-    // If you are using multiple modules, make sure to define both the port
-    // number and the module. For example you with a rangefinder:
-    // public static final int rangefinderPort = 1;
-    // public static final int rangefinderModule = 1;
-    
-    // TODO: Change the values below.
-    public static final int DRIVE_LEFT_MOTOR_1 = 0;
-    public static final int DRIVE_LEFT_MOTOR_2 = 0;
-    public static final int DRIVE_LEFT_MOTOR_3 = 0;
-    public static final int DRIVE_RIGHT_MOTOR_1 = 0;
-    public static final int DRIVE_RIGHT_MOTOR_2 = 0;
-    public static final int DRIVE_RIGHT_MOTOR_3 = 0;
-    public static final int PAD_DRIVER = 0;
-    public static final int ROLLER_MOTOR = 0;
-    public static final int ROLLER_DEPLOY = 0;
-}
